refactor: move type definitions to utils.types.ts
